Split native Zenmap scan progress helpers
